+++ test/unit/mongo_db_admin/process_mail.py
@@ -1,0 +1,176 @@
+#!/usr/bin/python
+# Classification (U)
+
+"""Program:  process_mail.py
+
+    Description:  Unit testing of process_mail in mongo_db_admin.py.
+
+    Usage:
+        test/unit/mongo_db_admin/process_mail.py
+
+    Arguments:
+
+"""
+
+# Libraries and Global Variables
+
+# Standard
+import sys
+import os
+
+if sys.version_info < (2, 7):
+    import unittest2 as unittest
+else:
+    import unittest
+
+# Third-party
+import json
+
+# Local
+sys.path.append(os.getcwd())
+import mongo_db_admin
+import version
+
+__version__ = version.__version__
+
+
+class Mail(object):
+
+    """Class:  Mail
+
+    Description:  Class stub holder for gen_class.Mail class.
+
+    Methods:
+        __init__
+        add_2_msg
+        send_mail
+
+    """
+
+    def __init__(self, lag_time=1):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+
+        """
+
+        self.lag_time = lag_time
+        self.data = None
+
+    def add_2_msg(self, data):
+
+        """Method:  add_2_msg
+
+        Description:  Stub method holder for Mail.add_2_msg.
+
+        Arguments:
+
+        """
+
+        self.data = data
+
+        return True
+
+    def send_mail(self):
+
+        """Method:  get_name
+
+        Description:  Stub method holder for Mail.send_mail.
+
+        Arguments:
+
+        """
+
+        return True
+
+
+class UnitTest(unittest.TestCase):
+
+    """Class:  UnitTest
+
+    Description:  Class which is a representation of a unit testing.
+
+    Methods:
+        setUp
+        test_mail_no_indent
+        test_str_to_email
+        test_dict_to_email
+        test_json_to_email
+
+    """
+
+    def setUp(self):
+
+        """Function:  setUp
+
+        Description:  Initialization for unit testing.
+
+        Arguments:
+
+        """
+
+        self.mail = Mail()
+        self.data = {"key": "value"}
+        self.data2 = "Key: Value"
+        self.indent = 4
+
+    def test_mail_no_indent(self):
+
+        """Function:  test_mail_no_indent
+
+        Description:  Test with dictionary to email with no indent passed.
+
+        Arguments:
+
+        """
+
+        self.assertFalse(mongo_db_admin.process_mail(self.mail, self.data))
+
+    def test_str_to_email(self):
+
+        """Function:  test_str_to_email
+
+        Description:  Test with string to email.
+
+        Arguments:
+
+        """
+
+        self.assertFalse(mongo_db_admin.process_mail(self.mail, self.data2,
+                                                     self.indent))
+
+    def test_dict_to_email(self):
+
+        """Function:  test_dict_to_email
+
+        Description:  Test with dictionary to email.
+
+        Arguments:
+
+        """
+
+        self.assertFalse(mongo_db_admin.process_mail(self.mail, self.data,
+                                                     self.indent))
+
+    def test_json_to_email(self):
+
+        """Function:  test_json_to_email
+
+        Description:  Test with JSON to email.
+
+        Arguments:
+
+        """
+
+        self.data = json.dumps(self.data, indent=4)
+
+        self.assertFalse(mongo_db_admin.process_mail(self.mail, self.data,
+                                                     self.indent))
+
+
+if __name__ == "__main__":
+    unittest.main()
+